test getting from interface Contract
